[fix] move test node to assets
