+++ tests/assets/node.ts
@@ -1,0 +1,50 @@
+const tableNode = {
+  type: 'table',
+  align: null,
+  children: [
+    {
+      type: 'tableRow',
+      children: [
+        {
+          type: 'tableCell',
+          children: [{type: 'text', value: 'foo'}]
+        },
+        {
+          type: 'tableCell',
+          children: [{type: 'text', value: 'bar'}]
+        }
+      ]
+    },
+    {
+      type: 'tableRow',
+      children: [
+        {
+          type: 'tableCell',
+          children: [{type: 'text', value: 'baz'}]
+        },
+        {
+          type: 'tableCell',
+          children: [{type: 'text', value: 'qux'}]
+        }
+      ]
+    }
+  ]
+}
+
+const listNode = {
+  type: 'list',
+  ordered: true,
+  start: 1,
+  spread: false,
+  children: [{
+    type: 'listItem',
+    checked: true,
+    spread: false,
+    children: [{
+      type: 'paragraph',
+      children: [{type: 'text', value: 'foo'}]
+    }]
+  }]
+}
+
+export { tableNode, listNode }
