+++ tests/compiler.test.ts
@@ -1,4 +1,5 @@
 import { compiler } from '../src/compiler'
+import { tableNode, listNode } from './assets/node'
 
 test('text', () => {
   const node = { type: 'text', value: 'Alpha bravo charlie.' }
@@ -51,52 +52,3 @@ test('table', () => {
 test('list', () => {
   expect(compiler['table'](listNode)).toBe('+ [x] foo')
 })
-
-const tableNode = {
-  type: 'table',
-  align: null,
-  children: [
-    {
-      type: 'tableRow',
-      children: [
-        {
-          type: 'tableCell',
-          children: [{type: 'text', value: 'foo'}]
-        },
-        {
-          type: 'tableCell',
-          children: [{type: 'text', value: 'bar'}]
-        }
-      ]
-    },
-    {
-      type: 'tableRow',
-      children: [
-        {
-          type: 'tableCell',
-          children: [{type: 'text', value: 'baz'}]
-        },
-        {
-          type: 'tableCell',
-          children: [{type: 'text', value: 'qux'}]
-        }
-      ]
-    }
-  ]
-}
-
-const listNode = {
-  type: 'list',
-  ordered: true,
-  start: 1,
-  spread: false,
-  children: [{
-    type: 'listItem',
-    checked: true,
-    spread: false,
-    children: [{
-      type: 'paragraph',
-      children: [{type: 'text', value: 'foo'}]
-    }]
-  }]
-}
